% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_UI.R
\name{incidence_ui}
\alias{incidence_ui}
\title{Web interface for incidence}
\usage{
incidence_ui()
}
\description{
This function opens up an application in a web browser for an interactive use
of RECON's incidence package.
}
\seealso{
The incidence package: \url{http://www.repidemicsconsortium.org/incidence/}
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
